requirements.txt
wheel
PyNaCl==1.4.0
discord.py==1.7.3
discord.py[voice]==1.7.3
pip
python-dotenv
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
idna==3.3
multidict==6.0.2
numpy==1.22.1
pycparser==2.21
six==1.16.0
timedelta==2020.12.3
typing-extensions==4.0.1
yarl==1.7.2
